StableOrderingFilter only for NamedModelViewSet
https://pulp.plan.io/issues/6347
fixes #6347
