+++ CHANGES/plugin_api/6347.feature
@@ -1,0 +1,2 @@
+`NamedModelViewSet` uses a default ordering of `-pulp_created` using the `StableOrderingFilter`.
+Users using the `ordering` keyword will be the primary ordering used when specified.
